+++ services/resources-getter.js
@@ -13,11 +13,6 @@ function ResourcesGetter(model, opts, params) {
   var segmentScope;
   var segmentWhere;
 
-  var fields = _.clone(schema.fields);
-  var associations = _.clone(model.associations);
-  var hasSearchFields = schema.searchFields && _.isArray(schema.searchFields);
-  var searchAssociationFields;
-
   var fieldNamesRequested = (function() {
     if (!params.fields || !params.fields[model.name]) { return null; }
 
@@ -41,41 +36,6 @@ function ResourcesGetter(model, opts, params) {
     return _.union(primaryKeyArray, params.fields[model.name].split(','),
       associationsForQuery);
   })();
-
-  function selectSearchFields() {
-    var searchFields = _.clone(schema.searchFields);
-    searchAssociationFields = _.remove(searchFields, function (field) {
-      return field.indexOf('.') !== -1;
-    });
-
-    _.remove(fields, function (field) {
-      return !_.includes(schema.searchFields, field.field);
-    });
-
-    var searchAssociationNames = _.map(searchAssociationFields,
-      function (association) { return association.split('.')[0]; });
-    associations = _.pick(associations, searchAssociationNames);
-
-    // NOTICE: Compute warnings to help developers to configure the
-    //         searchFields.
-    var fieldsSimpleNotFound = _.xor(searchFields,
-      _.map(fields, function (field) { return field.field; }));
-    var fieldsAssociationNotFound = _.xor(
-      _.map(searchAssociationFields, function (association) {
-        return association.split('.')[0];
-      }), _.keys(associations));
-
-    if (fieldsSimpleNotFound.length) {
-      Interface.logger.warn('Cannot find the fields [' + fieldsSimpleNotFound +
-        '] while searching records in model ' + model.name + '.');
-    }
-
-    if (fieldsAssociationNotFound.length) {
-      Interface.logger.warn('Cannot find the associations [' +
-        fieldsAssociationNotFound + '] while searching records in model ' +
-        model.name + '.');
-    }
-  }
 
   function handleFilterParams() {
     var where = {};
@@ -118,10 +78,6 @@ function ResourcesGetter(model, opts, params) {
   }
 
   function getAndCountRecords() {
-    if (hasSearchFields) {
-      selectSearchFields();
-    }
-
     var countOpts = {
       include: queryBuilder.getIncludes(model, fieldNamesRequested),
       where: getWhere()